# TL/DR:

When profiling the code I noticed hundreds of threads were created, making profiling not practical and maybe leading to some performance overhead. I then attempted to make a simple threadpool inside GGML to see if it made any difference at all. Gains at small scale are not notable, and inference does not change much. However, the more threads the more overhead and more importantly, the code is easier to profile (not 100s of threads to look at).

This is my first attempt at hacking the code, there might be cleaner attempts or better ways, I have seen PR #710, #851 which are similar. I'm also not sure of how to properly benchmark this, did launch `./bin/benchmark`.

## Perf on x86-64 (Genoa)

- One socket:
   - ./bin/benchmark -t 96 -i 100
      - without: (last line) average flops 1497.53
      - with : (last line) average flops 2970.68
- Two sockets:
   - ./bin/benchmark -t 192 -i 100
      - without: (last line) average flops 1044.79
      - with : (last line) average flops 3488.96


## Perf on x86-64 (Ivy Bridge)

- One socket:
   - ./bin/benchmark -t 18 -i 100
      - without: (last line) average flops 367.60
      - with : (last line) average flops 378.21
- Two sockets:
   - ./bin/benchmark -t 36 -i 100
      - without: (last line) average flops 633.22
      - with : (last line) average flops 706.8