@slaren 2024-05-17
You can use `llama-bench` to test the overall performance. I ran a test with and without KV offload since this is one case where the threads are started multiple times in each evaluation.

`LLAMA_CUDA=1 scripts/compare-commits.sh master thread_pool -nkvo 0,1`

| GPU         | Model         |   NKVO | Test        |   t/s master |   t/s thread_pool |   Speedup |
|:------------|:--------------|-------:|:------------|-------------:|------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 |      0 | pp512       |      4313.72 |           4525.64 |      1.05 |
| RTX 3090 Ti | llama 7B Q4_0 |      0 | tg128       |       154.96 |            154.17 |      0.99 |
| RTX 3090 Ti | llama 7B Q4_0 |      0 | pp512+tg128 |       622.75 |            625.94 |      1.01 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | pp512       |       619.01 |            527.14 |      0.85 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | tg128       |        21.55 |             49.80 |      2.31 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | pp512+tg128 |        84.56 |            135.97 |      1.61 |