@besnardjb 2024-05-17
Thanks for the pointer to `llama-bench`! sorry for missing it. I did run a few CPU tests on a bi-socket machine. I was not able to use the convenient `compare-commits.sh` script yet, as there seems to be some warning output piped in SQL.

Overall, there seem to be punctual gains, but it's close to measurement noise. I've also cleaned up the code a bit and tried to fix the Windows build ( :warning: without access to a Windows machine, unfortunately). For Windows, as apparently there is no direct equivalent of a semaphore, I have used busy waiting for work in threads to skip dependencies (measured its impact below on Linux too, just to see what it does).

Ran: `./bin/llama-bench -m ../models/7B/mistral-7b-instruct-v0.2.Q4_0.gguf -nkvo 0,1`

## ThreadPool (Semaphore)

```
| model                          |       size |     params | backend    |    threads |          nkvo |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ------------: | ---------------: |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         pp512 |    248.15 ± 1.38 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         tg128 |     22.39 ± 0.20 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |   pp512+tg128 |     81.91 ± 1.01 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         pp512 |    211.54 ± 8.78 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         tg128 |     19.73 ± 0.17 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |   pp512+tg128 |     70.52 ± 0.57 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         pp512 |   241.72 ± 10.81 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         tg128 |     22.25 ± 0.45 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |   pp512+tg128 |     82.73 ± 0.39 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         pp512 |    227.08 ± 2.27 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         tg128 |     19.42 ± 0.14 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |   pp512+tg128 |     70.71 ± 0.77 |
```


## Threadpool (ACTIVE Poll -- as for windows measured on Linux)

```
| model                          |       size |     params | backend    |    threads |          nkvo |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ------------: | ---------------: |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         pp512 |    246.28 ± 0.34 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         tg128 |     24.11 ± 0.10 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |   pp512+tg128 |     84.40 ± 0.54 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         pp512 |    224.57 ± 0.51 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         tg128 |     19.38 ± 0.23 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |   pp512+tg128 |     71.38 ± 0.79 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         pp512 |    247.67 ± 1.07 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         tg128 |     24.73 ± 0.48 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |   pp512+tg128 |     83.10 ± 0.75 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         pp512 |    236.53 ± 1.82 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         tg128 |     19.02 ± 0.08 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |   pp512+tg128 |     69.73 ± 0.14 |
```


# Master (9ca4abed893685692f90413e4d43153af12342d9)

```
| model                          |       size |     params | backend    |    threads |          nkvo |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | ---------: | ------------: | ------------: | ---------------: |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         pp512 |   246.89 ± 11.27 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |         tg128 |     19.84 ± 0.01 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             0 |   pp512+tg128 |     75.51 ± 0.43 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         pp512 |    238.93 ± 6.04 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |         tg128 |     18.17 ± 0.15 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             0 |   pp512+tg128 |     69.23 ± 0.48 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         pp512 |   253.94 ± 14.74 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |         tg128 |     19.85 ± 0.02 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |         96 |             1 |   pp512+tg128 |     75.34 ± 1.20 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         pp512 |   236.17 ± 16.98 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |         tg128 |     18.25 ± 0.08 |
| llama 7B Q4_0                  |   3.83 GiB |     7.24 B | CPU        |        192 |             1 |   pp512+tg128 |     69.36 ± 0.58 |
```