@besnardjb 2024-05-20
After looking into the Futex approach I realized it was about to add a lot of ifdefs and several lines of (relatively complex) codes. I then thought that I was in fact trying to implement an OpenMP parallel region. And thus, resorted to trying OpenMP directly. This basically translated in a single `#pragma omp` and avoids manual handling. In addition, what is nice is that it could be used on other for loops too elsewhere in the code without repeating the whole threadpool thingy (the runtime does it for free). I think it ends up simpler, but all comments are welcome of course.

I hope it works well on windows, lets see. 

Also I realized performance counters in `ggml_graph_compute` were not accounting for thread creation cost, I thus moved them a few line up on master to do the same measure in the two case.


## 72 Cores  (moved timings for thread creation)

| CPU                                       | Model         | Test        |   t/s master |   t/s openmp |   Speedup |
|:------------------------------------------|:--------------|:------------|-------------:|-------------:|----------:|
| Intel(R) Xeon(R) CPU E5-2697 v4 @ 2.30GHz | llama 7B Q4_0 | pp512       |        47.17 |        58.55 |      1.24 |
| Intel(R) Xeon(R) CPU E5-2697 v4 @ 2.30GHz | llama 7B Q4_0 | tg128       |         9.08 |        13.96 |      1.54 |
| Intel(R) Xeon(R) CPU E5-2697 v4 @ 2.30GHz | llama 7B Q4_0 | pp512+tg128 |        27.37 |        35.15 |      1.28 |

## 192 cores (moved timings for thread creation)

| CPU                             | Model         | Test        |   t/s master |   t/s openmp |   Speedup |
|:--------------------------------|:--------------|:------------|-------------:|-------------:|----------:|
| AMD EPYC 9654 96-Core Processor | llama 7B Q4_0 | pp512       |       226.86 |       241.64 |      1.07 |
| AMD EPYC 9654 96-Core Processor | llama 7B Q4_0 | tg128       |        18.24 |        21.42 |      1.17 |
| AMD EPYC 9654 96-Core Processor | llama 7B Q4_0 | pp512+tg128 |        69.03 |        76.94 |      1.11 |

# Old Laptop 4 cores (moved timings for thread creation)
| CPU                                       | Model          | Test        |   t/s master |   t/s openmp |   Speedup |
|:------------------------------------------|:---------------|:------------|-------------:|-------------:|----------:|
| Intel(R) Core(TM) i7-4712HQ CPU @ 2.30GHz | phi3 3B Q4_K_M | pp512       |        13.07 |        13.28 |      1.02 |
| Intel(R) Core(TM) i7-4712HQ CPU @ 2.30GHz | phi3 3B Q4_K_M | tg128       |         7.51 |         7.60 |      1.01 |
| Intel(R) Core(TM) i7-4712HQ CPU @ 2.30GHz | phi3 3B Q4_K_M | pp512+tg128 |        11.38 |        11.28 |      0.99 |


# H100 (moved timings for thread creation)

| GPU              | Model         |   NKVO | Test        |   t/s master |   t/s openmp |   Speedup |
|:-----------------|:--------------|-------:|:------------|-------------:|-------------:|----------:|
| NVIDIA H100 PCIe | llama 7B Q4_0 |      0 | pp512       |      6366.18 |      6222.38 |      0.98 |
| NVIDIA H100 PCIe | llama 7B Q4_0 |      0 | tg128       |       130.68 |       120.55 |      0.92 |
| NVIDIA H100 PCIe | llama 7B Q4_0 |      0 | pp512+tg128 |       549.15 |       515.65 |      0.94 |
| NVIDIA H100 PCIe | llama 7B Q4_0 |      1 | pp512       |       601.44 |       711.63 |      1.18 |
| NVIDIA H100 PCIe | llama 7B Q4_0 |      1 | tg128       |         4.94 |        33.61 |      6.81 |
| NVIDIA H100 PCIe | llama 7B Q4_0 |      1 | pp512+tg128 |        22.21 |       128.38 |      5.78 |