@slaren 2024-05-20
Unfortunately the omp is slower for me, and it causes a significant performance regression without nkvo.

| GPU         | Model         |   NKVO | Test        |   t/s master |   t/s thread_pool |   Speedup |
|:------------|:--------------|-------:|:------------|-------------:|------------------:|----------:|
| RTX 3090 Ti | llama 7B Q4_0 |      0 | pp512       |      4512.80 |           4076.89 |      0.90 |
| RTX 3090 Ti | llama 7B Q4_0 |      0 | tg128       |       154.23 |            140.59 |      0.91 |
| RTX 3090 Ti | llama 7B Q4_0 |      0 | pp512+tg128 |       627.11 |            570.19 |      0.91 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | pp512       |       526.36 |            489.66 |      0.93 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | tg128       |        21.86 |             42.83 |      1.96 |
| RTX 3090 Ti | llama 7B Q4_0 |      1 | pp512+tg128 |        84.62 |            131.32 |      1.55 |